Rename to gqlclient
