+++ internal/gqlclient/error.go
@@ -1,4 +1,4 @@
-package graphql
+package gqlclient
 
 import "fmt"
