+++ internal/gqlclient/graphql.go
@@ -29,7 +29,7 @@
 //
 //	httpclient := &http.Client{}
 //	client := graphql.NewClient("https://machinebox.io/graphql", graphql.WithHTTPClient(httpclient))
-package graphql
+package gqlclient
 
 import (
 	"bytes"
