+++ internal/gqlclient/graphql_multipart_test.go
@@ -1,4 +1,4 @@
-package graphql
+package gqlclient
 
 import (
 	"context"
