+++ pkg/platform/api/vulnerabilities/vulnerabilities.go
@@ -1,13 +1,13 @@
 package vulnerabilities
 
 import (
-	"github.com/ActiveState/cli/internal/graphql"
+	"github.com/ActiveState/cli/internal/gqlclient"
 	"github.com/ActiveState/cli/pkg/platform/api"
 	"github.com/ActiveState/cli/pkg/platform/authentication"
 )
 
-func New(auth *authentication.Auth) *graphql.Client {
-	client := graphql.New(api.GetServiceURL(api.ServiceVulnerabilities).String(), 0)
+func New(auth *authentication.Auth) *gqlclient.Client {
+	client := gqlclient.New(api.GetServiceURL(api.ServiceVulnerabilities).String(), 0)
 
 	// Most requests to this service require authentication
 	if auth != nil && auth.Authenticated() {
