+++ pkg/platform/model/buildplanner/buildplanner.go
@@ -3,7 +3,7 @@ package buildplanner
 import (
 	"time"
 
-	"github.com/ActiveState/cli/internal/graphql"
+	"github.com/ActiveState/cli/internal/gqlclient"
 	"github.com/ActiveState/cli/internal/logging"
 	"github.com/ActiveState/cli/pkg/platform/api"
 	"github.com/ActiveState/cli/pkg/platform/authentication"
@@ -12,7 +12,7 @@ import (
 const clientDeprecationErrorKey = "CLIENT_DEPRECATION_ERROR"
 
 type client struct {
-	gqlClient *graphql.Client
+	gqlClient *gqlclient.Client
 }
 
 type BuildPlanner struct {
@@ -40,7 +40,7 @@ func NewBuildPlannerModel(auth *authentication.Auth, cache cacher) *BuildPlanner
 	bpURL := api.GetServiceURL(api.ServiceBuildPlanner).String()
 	logging.Debug("Using build planner at: %s", bpURL)
 
-	gqlClient := graphql.NewWithOpts(bpURL, 0, graphql.WithHTTPClient(api.NewHTTPClient()))
+	gqlClient := gqlclient.NewWithOpts(bpURL, 0, gqlclient.WithHTTPClient(api.NewHTTPClient()))
 
 	if auth != nil && auth.Authenticated() {
 		gqlClient.SetTokenProvider(auth)
